+++ game_class.py
@@ -1,10 +1,8 @@
 import pygame
-import sys
 import texture_class
 import level_class
-import gui
 import renderer
-from pygame.locals import QUIT
+import event_class
 
 
 class game:
@@ -26,8 +24,8 @@ class game:
         # manager inicializers
         self.texture_manager = texture_class.Texture_manager()
         self.renderer = renderer.Render_manager(self)
-        self.gui_manager = gui.GUI_manager(self)
         self.level_manager = level_class.Level_manager(self)
+        self.event_manager = event_class.Event_manager(self)
 
         # inicialization of the first level
         self.level_manager.set_level("GUI_mainMenu.json")
@@ -38,30 +36,8 @@ class game:
         # this is the player that is currently in the game
         self.local_player = None
 
-    # used to render the game frame and update the fps
-    def fps_render(self):
-        self.renderer.render()
-        self.delta_time = self.__clock__.tick(self.FPS)
-
-        pygame.display.update()
-        # pygame.display.flip()
-
-    # used to process all the events that happen in the game TODO: create event manager to manage this
-    def event_loop(self):
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                pygame.quit()
-                sys.exit()
-            self.gui_manager.process_events(event)
-
-        key_pressed = pygame.key.get_pressed()
-        for i in self.trigger_list:
-            i.interaction(key=key_pressed, interaction_author=self.local_player)
-
     # the default game loop, it is called every frame to process everything
     def game_loop(self):
         while (True):
-            # TODO: add a better way of showing fps count
-            # print(self.__clock__.get_fps())
-            self.event_loop()
-            self.fps_render()
+            self.event_manager.process_events()
+            self.renderer.render()
